requirements.txt
PySide6
opencv-python
ultralytics
numpy
mysql-connector-python
matplotlib
